Add files via upload
The IR is used for starting the program. If the IR sensor detects the object, then the weight of that particular object is calculated using Weight sensor and sent to the System. Following this, the camera gets started and the object is captured twice - one with the name and brand and the other with the expiration date, manufacturing date and price. 
These are sent to the GCP Bigquery and are stored. Which can be edited, removed, and the items can be manually added through the website.
If a particular item is searched in the inventory website, then the total number of item with that name or brand is displayed with the total count.
Finally, the freshness of the object is calculated in 2 ways.
- For products, it is calculated through the expiry date
- For Fruits and Vegetables, it is calculated through the color from many parts of the image and advanced algorithms using GCP. It return either Fresh or Rotten as shown at the end of the video simulation.